use int for ticks

+++ src/my_monst.hpp
@@ -199,47 +199,47 @@ public:
   //
   // Avoiding numeric_limits to reduce the number of headers to pull in.
   //
-  uint32_t tick_last_did_something {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_did_something {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // When this thing was dropped
   //
-  uint32_t tick_last_dropped {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_dropped {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // To prevent teleport looping
   //
-  uint32_t tick_last_teleported {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_teleported {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // If things have stopped moving, perform location checks on where theuy
   // are now. This handles things like shoving a monst into a chasm. We do
   // location checks on the ends of moves, but this is a backup and will
   // also handle things that do not move, like a staff that is now on fire.
   //
-  uint32_t tick_last_location_check {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_location_check {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // Don't try and escape too often else the player can never kill the monst!
   //
-  uint32_t tick_last_escape {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
-  uint32_t tick_last_level_change {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_escape {-1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_level_change {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // When we were last attacked
   //
-  uint32_t tick_last_i_was_attacked {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_i_was_attacked {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // Last time we had to hold our breath
   //
-  uint32_t tick_last_poison_gas_exposure {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_poison_gas_exposure {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // When we last attacked
   //
-  uint32_t tick_last_i_attacked {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_i_attacked {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // When we last tried to attack and missed or succeeded
   //
-  uint32_t tick_last_i_tried_to_attack {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_i_tried_to_attack {-1 /* std::numeric_limits< uint32_t >::max() */};
   //
   // This is to ensure things do not wake and immediately attack
   //
-  uint32_t tick_last_awoke {(uint32_t) -1 /* std::numeric_limits< uint32_t >::max() */};
+  int tick_last_awoke {-1 /* std::numeric_limits< uint32_t >::max() */};
 
   //
   // List of things I own:
